+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added keybinding to toggle vertical synchronization
 - Round displayed FPS to an integral value
 - Fix orbital plane of various moons
+- Fix rotation direction of Uranus
 
 # v0.3.0 - 2019-08-17
 - Use Moon orbital elements of 2019-08-11